organize and clean service logics

+++ post-service/src/main/java/com/bfsforum/postservice/controller/PostController.java
@@ -135,6 +135,10 @@ public class PostController {
 		boolean isOwner = post.getUserId().equals(userId);
 		boolean isAdmin = role.equalsIgnoreCase("ADMIN");
 		
+		if (!isOwner && !isAdmin) {
+			throw new UnauthorizedException("You can only delete your own posts");
+		}
+		
 		postService.deletePost(id);
 		return ResponseEntity.noContent().build();
 	}
@@ -203,61 +207,80 @@ public class PostController {
 		return ResponseEntity.ok(posts);
 	}
 	
-	// admin
-	@PutMapping("/{id}/ban")
-	public ResponseEntity<Post> banPost(@PathVariable String id) {
-		
-		Post post = postService.getPostById(id)
-				.orElseThrow(() -> new PostNotFoundException(id));
-		
-		post.setStatus(PostStatus.BANNED);
-		Post updatedPost = postService.updatePost(id, post);
+	@PutMapping("/{id}/status")
+	public ResponseEntity<Post> updatePostStatus(
+			@PathVariable String id,
+			@RequestParam String status,
+			@RequestHeader("User-Id") Long userId,
+			@RequestHeader(value = "Role", defaultValue = "USER") String role){
+	
+		Post updatedPost = postService.updatePostStatus(id, status, userId, role);
 		return ResponseEntity.ok(updatedPost);
 	}
 	
-	@PutMapping("/{id}/unban")
-	public ResponseEntity<Post> unbanPost(@PathVariable String id) {
-		Post post = postService.getPostById(id)
-				.orElseThrow(() -> new PostNotFoundException(id));
-		post.setStatus(PostStatus.PUBLISHED);
-		Post updatedPost = postService.updatePost(id, post);
-		return ResponseEntity.ok(updatedPost);
-	}
-	
-	@PutMapping("/{id}/archive")
-	public ResponseEntity<Post> archivePost(@PathVariable String id) {
-		Post post = postService.getPostById(id)
-				.orElseThrow(() -> new PostNotFoundException(id));
-		post.setIsArchived(true);
-		Post updatedPost = postService.updatePost(id, post);
-		return ResponseEntity.ok(updatedPost);
-	}
-	
-	@PutMapping("/{id}/unarchive")
-	public ResponseEntity<Post> unarchivePost(@PathVariable String id) {
-		Post post = postService.getPostById(id)
-				.orElseThrow(() -> new PostNotFoundException(id));
-		
-		post.setIsArchived(false);
-		Post updatedPost = postService.updatePost(id, post);
-		return ResponseEntity.ok(updatedPost);
-	}
-	
-	@PutMapping("/{id}/hide")
-	public ResponseEntity<Post> hidePost(@PathVariable String id) {
-		Post post = postService.getPostById(id)
-				.orElseThrow(() -> new PostNotFoundException(id));
-		post.setStatus(PostStatus.HIDDEN);
-		Post updatedPost = postService.updatePost(id, post);
-		return ResponseEntity.ok(updatedPost);
-	}
-	
-	@PutMapping("/{id}/unhide")
-	public ResponseEntity<Post> unhidePost(@PathVariable String id) {
-		Post post = postService.getPostById(id)
-				.orElseThrow(() -> new PostNotFoundException(id));
-		post.setStatus(PostStatus.PUBLISHED);
-		Post updatedPost = postService.updatePost(id, post);
-		return ResponseEntity.ok(updatedPost);
-	}
+//	// admin
+//	@PutMapping("/{id}/ban")
+//	public ResponseEntity<Post> banPost(@PathVariable String id) {
+//
+//		Post post = postService.getPostById(id)
+//				.orElseThrow(() -> new PostNotFoundException(id));
+//
+//		post.setStatus(PostStatus.BANNED);
+//		Post updatedPost = postService.updatePost(id, post);
+//		return ResponseEntity.ok(updatedPost);
+//	}
+//
+//	@PutMapping("/{id}/unban")
+//	public ResponseEntity<Post> unbanPost(@PathVariable String id) {
+//		Post post = postService.getPostById(id)
+//				.orElseThrow(() -> new PostNotFoundException(id));
+//		post.setStatus(PostStatus.PUBLISHED);
+//		Post updatedPost = postService.updatePost(id, post);
+//		return ResponseEntity.ok(updatedPost);
+//	}
+//
+//	@PutMapping("/{id}/archive")
+//	public ResponseEntity<Post> archivePost(@PathVariable String id,
+//	                                        @RequestHeader("User-Id") Long userId,
+//	                                        @RequestHeader(value = "Role", defaultValue = "USER") String role) {
+//		Post post = postService.getPostById(id)
+//				.orElseThrow(() -> new PostNotFoundException(id));
+//		post.setIsArchived(true);
+//		Post updatedPost = postService.updatePost(id, post);
+//		return ResponseEntity.ok(updatedPost);
+//	}
+//
+//	@PutMapping("/{id}/unarchive")
+//	public ResponseEntity<Post> unarchivePost(@PathVariable String id,
+//	                                          @RequestHeader("User-Id") Long userId,
+//	                                          @RequestHeader(value = "Role", defaultValue = "USER") String role) {
+//		Post post = postService.getPostById(id)
+//				.orElseThrow(() -> new PostNotFoundException(id));
+//
+//		post.setIsArchived(false);
+//		Post updatedPost = postService.updatePost(id, post);
+//		return ResponseEntity.ok(updatedPost);
+//	}
+//
+//	@PutMapping("/{id}/hide")
+//	public ResponseEntity<Post> hidePost(@PathVariable String id,
+//	                                     @RequestHeader("User-Id") Long userId,
+//	                                     @RequestHeader(value = "Role", defaultValue = "USER") String role) {
+//		Post post = postService.getPostById(id)
+//				.orElseThrow(() -> new PostNotFoundException(id));
+//		post.setStatus(PostStatus.HIDDEN);
+//		Post updatedPost = postService.updatePost(id, post);
+//		return ResponseEntity.ok(updatedPost);
+//	}
+//
+//	@PutMapping("/{id}/unhide")
+//	public ResponseEntity<Post> unhidePost(@PathVariable String id,
+//	                                       @RequestHeader("User-Id") Long userId,
+//	                                       @RequestHeader(value = "Role", defaultValue = "USER") String role) {
+//		Post post = postService.getPostById(id)
+//				.orElseThrow(() -> new PostNotFoundException(id));
+//		post.setStatus(PostStatus.PUBLISHED);
+//		Post updatedPost = postService.updatePost(id, post);
+//		return ResponseEntity.ok(updatedPost);
+//	}
 }
